apps/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "AlexanderMatveev2908",email = "matveevalexander470@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "attrs (>=25.3.0,<26.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "regex (>=2025.7.34,<2026.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.42,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "aioboto3 (>=15.0.0,<16.0.0)",
    "types-aiobotocore-s3 (>=2.23.2,<3.0.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
pyright = "^1.1.403"
watchfiles = "^1.1.0"
black = "^25.1.0"
ruff = "^0.12.7"
tomlkit = "^0.13.3"
sqlalchemy = {extras = ["mypy"], version = "^2.0.42"}
types-aiofiles = "^24.1.0.20250801"
pytest-json-report = "^1.5.0"


[[tool.poetry.packages]]
include = "src"
from = "."

[tool.black]
line-length = 79
target-version = ["py312"]

[tool.flake8]
ignore = ["E203"]
max-line-length = 79

[tool.ruff]
line-length = 79

[tool.ruff.lint]
ignore = ["E203", "E741"]
[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["*_t", "test_*"]
python_classes = ["*TST", "Test*"]
addopts = "-v --no-header --json-report --json-report-file=test-results/reports.json"
asyncio_mode = "strict"
